// packages/porygon/src/crud/models/domain/CrudModuleType.ts
export enum CrudModuleType {
  createEntityAdapter = 'createEntityAdapter',
  createEntityInteractor = 'createEntityInteractor',
  deleteEntityAdapter = 'deleteEntityAdapter',
  deleteEntityInteractor = 'deleteEntityInteractor',
  readEntityAdapter = 'readEntityAdapter',
  readManyEntityInteractor = 'readManyEntityInteractor',
  readOneEntityInteractor = 'readOneEntityInteractor',
  updateEntityAdapter = 'updateEntityAdapter',
  updateEntityInteractor = 'updateEntityInteractor',
}